LibHTTP: Unify and generalise response handling logic
